remove bespoke enabled logic, refactor to fix cyclic import
