+++ pilot/pkg/networking/serviceentry/ip_autoallocation.go
